+++ src/dipdup/cli.py
@@ -3,8 +3,8 @@ import asyncio
 import atexit
 import logging
 import sys
-from collections.abc import Awaitable
 from collections.abc import Callable
+from collections.abc import Coroutine
 from contextlib import AsyncExitStack
 from contextlib import suppress
 from dataclasses import dataclass
@@ -15,7 +15,8 @@ from typing import Any
 from typing import TypeVar
 from typing import cast
 
-import asyncclick as click
+import click
+import uvloop
 
 from dipdup import __version__
 from dipdup import env
@@ -59,13 +60,6 @@ NO_CONFIG_CMDS = {
     'migrate',
     'config',
 }
-# NOTE: Our signal handler conflicts with Click's one in prompt mode
-NO_SIGNALS_CMDS = {
-    *NO_CONFIG_CMDS,
-    None,
-    'schema',
-    'wipe',
-}
 
 
 _logger = logging.getLogger(__name__)
@@ -103,7 +97,7 @@ def _print_help_atexit(error: Exception, report_id: str) -> None:
     atexit.register(_print)
 
 
-WrappedCommandT = TypeVar('WrappedCommandT', bound=Callable[..., Awaitable[None]])
+WrappedCommandT = TypeVar('WrappedCommandT', bound=Callable[..., Coroutine[Any, Any, None]])
 
 
 @dataclass
@@ -114,12 +108,9 @@ class CLIContext:
 
 def _cli_wrapper(fn: WrappedCommandT) -> WrappedCommandT:
     @wraps(fn)
-    async def wrapper(ctx: click.Context, *args: Any, **kwargs: Any) -> None:
-        signals = ctx.invoked_subcommand not in NO_SIGNALS_CMDS
-        set_up_process(signals)
-
+    def wrapper(ctx: click.Context, *args: Any, **kwargs: Any) -> None:
         try:
-            await fn(ctx, *args, **kwargs)
+            uvloop.run(fn(ctx, *args, **kwargs))
         except (KeyboardInterrupt, asyncio.CancelledError):
             pass
         except Exception as e:
@@ -134,6 +125,10 @@ def _cli_wrapper(fn: WrappedCommandT) -> WrappedCommandT:
             save_report(package, None)
 
     return cast(WrappedCommandT, wrapper)
+
+
+def _cli_unwrapper(cmd: click.Command) -> Callable[..., Coroutine[Any, Any, None]]:
+    return cmd.callback.__wrapped__.__wrapped__  # type: ignore[no-any-return,union-attr]
 
 
 async def _check_version() -> None:
@@ -209,6 +204,8 @@ def _skip_cli_group() -> bool:
 @click.pass_context
 @_cli_wrapper
 async def cli(ctx: click.Context, config: list[str], env_file: list[str]) -> None:
+    set_up_process()
+
     if _skip_cli_group():
         return
 
@@ -360,7 +357,11 @@ async def migrate(ctx: click.Context, dry_run: bool) -> None:
         config_paths=ctx.parent.params['config'],
         config=config,
     )
-    await ctx.invoke(init, base=True, force=True)
+    await _cli_unwrapper(init)(
+        ctx=ctx,
+        base=True,
+        force=True,
+    )
 
 
 @cli.group()
@@ -577,7 +578,7 @@ async def schema_wipe(ctx: click.Context, immune: bool, force: bool) -> None:
 
     if not force:
         try:
-            assert sys.__stdin__.isatty()
+            assert sys.__stdin__.isatty()  # type: ignore[union-attr]
             click.confirm(
                 f"You're about to wipe schema `{url}`. All indexed data will be irreversibly lost, are you sure?",
                 abort=True,
@@ -710,7 +711,6 @@ async def new(
     template: str | None,
 ) -> None:
     """Create a new project interactively."""
-    import os
 
     from survey._widgets import Escape  # type: ignore[import-untyped]
 
@@ -740,9 +740,12 @@ async def new(
         config_paths=[Path(answers['package']).joinpath(ROOT_CONFIG).as_posix()],
         config=config,
     )
-    # NOTE: datamodel-codegen fails otherwise
-    os.chdir(answers['package'])
-    await ctx.invoke(init, base=True, force=force)
+    await _cli_unwrapper(init)(
+        ctx=ctx,
+        base=False,
+        force=force,
+        include=[],
+    )
 
     green_echo('Project created successfully!')
     green_echo(f"Enter `{answers['package']}` directory and see README.md for the next steps.")