+++ src/dipdup/config/tezos_operations.py
@@ -10,8 +10,9 @@ from pydantic.dataclasses import dataclass
 from pydantic.fields import Field
 
 from dipdup.config import Alias
-from dipdup.config import CodegenMixin
 from dipdup.config import HandlerConfig
+from dipdup.config._mixin import CodegenMixin
+from dipdup.config._mixin import SubgroupIndexMixin
 from dipdup.config.tezos import TezosContractConfig
 from dipdup.config.tezos import TezosIndexConfig
 from dipdup.config.tezos_tzkt import TezosTzktDatasourceConfig
@@ -19,6 +20,7 @@ from dipdup.exceptions import ConfigInitializationException
 from dipdup.exceptions import ConfigurationError
 from dipdup.models.tezos import TezosOperationType
 from dipdup.models.tezos_tzkt import OriginationSubscription
+from dipdup.models.tezos_tzkt import SmartRollupCementSubscription
 from dipdup.models.tezos_tzkt import SmartRollupExecuteSubscription
 from dipdup.models.tezos_tzkt import TransactionSubscription
 from dipdup.utils import pascal_to_snake
@@ -28,27 +30,6 @@ if TYPE_CHECKING:
     from collections.abc import Iterator
 
     from dipdup.subscriptions import Subscription
-
-
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
-class SubgroupIndexMixin:
-    """`subgroup_index` field to track index of operation in group
-
-    :param subgroup_index:
-    """
-
-    def __post_init__(self) -> None:
-        self._subgroup_index: int | None = None
-
-    @property
-    def subgroup_index(self) -> int:
-        if self._subgroup_index is None:
-            raise ConfigInitializationException
-        return self._subgroup_index
-
-    @subgroup_index.setter
-    def subgroup_index(self, value: int) -> None:
-        self._subgroup_index = value
 
 
 class TezosOperationsPatternConfig(CodegenMixin):
@@ -358,6 +339,16 @@ class TezosOperationsIndexConfig(TezosIndexConfig):
                         raise ConfigInitializationException
                     if contract_config.address and contract_config.address.startswith('sr1'):
                         subs.add(SmartRollupExecuteSubscription(address=contract_config.address))
+
+        if TezosOperationType.sr_cement in self.types:
+            if self.merge_subscriptions:
+                subs.add(SmartRollupCementSubscription())
+            else:
+                for contract_config in self.contracts:
+                    if not isinstance(contract_config, TezosContractConfig):
+                        raise ConfigInitializationException
+                    if contract_config.address and contract_config.address.startswith('sr1'):
+                        subs.add(SmartRollupCementSubscription(address=contract_config.address))
 
         return subs