+++ src/dipdup/indexes/evm_events/index.py
@@ -14,9 +14,9 @@ from dipdup.indexes.evm_events.fetcher import EvmNodeEventFetcher
 from dipdup.indexes.evm_events.fetcher import EvmSubsquidEventFetcher
 from dipdup.indexes.evm_events.matcher import match_events
 from dipdup.models import RollbackMessage
+from dipdup.models._subsquid import SubsquidMessageType
 from dipdup.models.evm import EvmEvent
 from dipdup.models.evm import EvmEventData
-from dipdup.models.subsquid import SubsquidMessageType
 from dipdup.prometheus import Metrics
 
 QueueItem = tuple[EvmEventData, ...] | RollbackMessage
@@ -90,10 +90,19 @@ class EvmEventsIndex(
         if not self.node_datasources:
             raise FrameworkException('Creating EvmNodeEventFetcher, but no `evm.node` datasources available')
 
+        addresses = set()
+        for handler_config in self._config.handlers:
+            if handler_config.contract.address:
+                addresses.add(handler_config.contract.address)
+            else:
+                addresses.clear()
+                break
+
         return EvmNodeEventFetcher(
             datasources=self.node_datasources,
             first_level=first_level,
             last_level=last_level,
+            addresses=addresses,
         )
 
     def _match_level_data(