+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+### Fixed
+
+- tezos.operations: Fixed `sr_cement` operation index subscription.
+
+### Other
+
+- deps: Use `uvloop` to improve asyncio performance.
+
+## [8.0.0b4] - 2024-07-20
+
 ### Added
 
 - config: Publish JSON schemas for config validation and autocompletion.
@@ -14,6 +24,14 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ### Fixed
 
+- evm.events: Improve fetching event batches from node.
+- models: Fixed `CachedModel` preloading.
+
+## [7.5.9] - 2024-07-20
+
+### Fixed
+
+- evm.events: Improve fetching event batches from node.
 - models: Fixed `CachedModel` preloading.
 
 ## [8.0.0b3] - 2024-07-04
@@ -1535,7 +1553,8 @@ This release contains no changes except for the version number.
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[Unreleased]: https://github.com/dipdup-io/dipdup/compare/8.0.0b3...HEAD
+[Unreleased]: https://github.com/dipdup-io/dipdup/compare/8.0.0b4...HEAD
+[8.0.0b4]: https://github.com/dipdup-io/dipdup/compare/8.0.0b3...8.0.0b4
 [8.0.0b3]: https://github.com/dipdup-io/dipdup/compare/8.0.0b2...8.0.0b3
 [7.5.8]: https://github.com/dipdup-io/dipdup/compare/7.5.7...7.5.8
 [8.0.0b2]: https://github.com/dipdup-io/dipdup/compare/8.0.0b1...8.0.0b2